[feat] config 수정
