+++ PwaFrontend/vite.config.ts
@@ -15,7 +15,7 @@ export default defineConfig(({ mode }) => {
     resolve: {
       alias: {
         '@': path.resolve(__dirname, './src'),
-        '@shared': path.resolve(__dirname, '../shared'),
+        '@shared': path.resolve(__dirname, '../shared')
       },
     },
     server: {
